pyproject.toml
[project]
name = "generate_fake_user_activity"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.1.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
]
